Update Swift packages to use the XCFramework from the matching channel

+++ packaging/swift/update-ice-swift-nightly.sh
@@ -15,13 +15,23 @@ if [ -z "${STAGING_DIR:-}" ]; then
     exit 1
 fi
 
+if [ -z "${CHANNEL:-}" ]; then
+    echo "Error: CHANNEL environment variable is not set"
+    exit 1
+fi
+
+if [ -z "${QUALITY:-}" ]; then
+    echo "Error: QUALITY environment variable is not set"
+    exit 1
+fi
+
 # Get Git repository root directory
 root_dir=$(git rev-parse --show-toplevel)
 cd "$root_dir/packaging/swift"
 
 # Clone the ice-swift-nightly repository and add the Ice for Swift sources
 [ -d ice-swift-nightly ] && rm -rf ice-swift-nightly
-git clone "https://x-access-token:${ICE_NIGHTLY_PUBLISH_TOKEN}@github.com/zeroc-ice/ice-swift-nightly.git" -b main
+git clone "https://x-access-token:${ICE_NIGHTLY_PUBLISH_TOKEN}@github.com/zeroc-ice/ice-swift-nightly.git" -b ${CHANNEL}
 cd ice-swift-nightly
 
 # Remove existing directories to avoid keeping stale files that may have been removed from the Ice repository.
@@ -33,13 +43,20 @@ cp -rfv ../../../slice .
 cp -rfv ../../../swift .
 cp -v ../../../Package.swift .
 
+# Compute the download path based on quality (matching publish-xcframework-packages.yml)
+if [ "${QUALITY}" = "stable" ]; then
+    download_path="ice/${CHANNEL}"
+else
+    download_path="ice/${QUALITY}/${CHANNEL}"
+fi
+
 # Update the Packages.swift file with the URL and Checksum for the xcframeworks
 for zip_file in "${STAGING_DIR}"/*.zip; do
     echo "Processing zip file: ${zip_file}"
     zip_name=$(basename -s ".xcframework.zip" "${zip_file}")
     name="${zip_name%%-*}"
     version="${zip_name#*-}"
-    zip_url="https://download.zeroc.com/ice/nightly/${zip_name}.xcframework.zip"
+    zip_url="https://download.zeroc.com/${download_path}/${zip_name}.xcframework.zip"
 
     checksum=$(shasum -a 256 "${zip_file}" | cut -d ' ' -f 1)
     indent=$(printf "%12s" "") # indentation for the checksum line
@@ -52,6 +69,6 @@ done
 git add .
 git config user.name "ZeroC"
 git config user.email "git@zeroc.com"
-git commit -m "ice: $version Nightly build"
-git tag -a "$version" -m "ice: $version"
-git push origin main --tags
+git commit -m "ice: ${version} ${QUALITY} build"
+git tag -a "${version}" -m "ice: ${version} ${QUALITY} build"
+git push origin ${CHANNEL} --tags
